+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added [#21](https://github.com/dartoos-dev/eo_color/issues/21)
+- Beautiful showcase application of the color swatches as example.
 
 ## [0.0.2] - 2021-05-15
 ### Added [#16](https://github.com/dartoos-dev/eo_color/issues/16)